Split Distribution.Version
